redactor/requirements.txt
aspose-words==22.11.0 ; python_version >= "3.9" and python_version < "3.11"
click==8.1.3 ; python_version >= "3.9" and python_version < "3.11"
colorama==0.4.6 ; python_version >= "3.9" and python_version < "3.11" and platform_system == "Windows"
flask==2.2.2 ; python_version >= "3.9" and python_version < "3.11"
flaskwebgui==1.0.3 ; python_version >= "3.9" and python_version < "3.11"
fpdf==1.7.2 ; python_version >= "3.9" and python_version < "3.11"
h11==0.14.0 ; python_version >= "3.9" and python_version < "3.11"
importlib-metadata==5.1.0 ; python_version >= "3.9" and python_version < "3.10"
itsdangerous==2.1.2 ; python_version >= "3.9" and python_version < "3.11"
jinja2==3.1.2 ; python_version >= "3.9" and python_version < "3.11"
markupsafe==2.1.1 ; python_version >= "3.9" and python_version < "3.11"
psutil==5.9.4 ; python_version >= "3.9" and python_version < "3.11"
pymupdf==1.18.9 ; python_version >= "3.9" and python_version < "3.11"
uvicorn==0.20.0 ; python_version >= "3.9" and python_version < "3.11"
werkzeug==2.2.2 ; python_version >= "3.9" and python_version < "3.11"
zipp==3.10.0 ; python_version >= "3.9" and python_version < "3.10"
